refactor(bpane-host): decompose cdp_scroll and scroll_emit into focused submodules
Split the two remaining large files into meaningful abstractions:

cdp_scroll.rs (496 lines) → 5 files:
- cdp_hints.rs (97): snapshot_cdp_hints() — drain events, read CDP state
- cdp_scroll_track.rs (53): track_cdp_scroll() — extract CDP scroll delta
- scroll_resolve.rs (199): resolve_scroll() — content detection, drift, cooldown
- scroll_trust.rs (148): arbitrate_scroll_trust() — CDP/content trust arbitration
- cdp_scroll.rs (72): thin orchestrator calling the above 3

scroll_emit.rs (466 lines) → 5 files:
- dirty_set.rs (55): build_dirty_set() — emit coords, XDamage filter
- scroll_partition.rs (107): partition_and_compare() — content/chrome split
- scroll_residual.rs (190): analyze_scroll_residual() — fallback/thin mode
- scroll_send.rs (141): send_scroll_frames() — ScrollCopy, GridOffset, stats
- scroll_emit.rs (54): thin orchestrator

tile_loop/ now has 18 files, all under 200 lines (except mod.rs at 266
for the struct definition). All 303 tests pass.
